// src/schema.ts
export interface paths {
// ...
  "/players/{account_id}/recentMatches": {
    /** Recent matches played */
    get: {
      parameters: {
        path: {
          /** Steam32 account ID */
          account_id: number;
        };
      };
      responses: {
        /** Success */
        200: {
          schema: {
            /** Match ID */
            match_id?: number;
            /** Which slot the player is in. 0-127 are Radiant, 128-255 are Dire */
            player_slot?: number;
            /** Boolean indicating whether Radiant won the match */
            radiant_win?: boolean;
            /** Duration of the game in seconds */
            duration?: number;
            /** Integer corresponding to game mode played. List of constants can be found here: https://github.com/odota/dotaconstants/blob/master/json/game_mode.json */
            game_mode?: number;
            /** Integer corresponding to lobby type of match. List of constants can be found here: https://github.com/odota/dotaconstants/blob/master/json/lobby_type.json */
            lobby_type?: number;
            /** The ID value of the hero played */
            hero_id?: number;
            /** Start time of the match in seconds elapsed since 1970 */
            start_time?: number;
            /** version */
            version?: number;
            /** Total kills the player had at the end of the match */
            kills?: number;
            /** Total deaths the player had at the end of the match */
            deaths?: number;
            /** Total assists the player had at the end of the match */
            assists?: number;
            /** Skill bracket assigned by Valve (Normal, High, Very High) */
            skill?: number;
            /** Integer corresponding to which lane the player laned in for the match */
            lane?: number;
            /** lane_role */
            lane_role?: number;
            /** Boolean describing whether or not the player roamed */
            is_roaming?: boolean;
            /** cluster */
            cluster?: number;
            /** Integer describing whether or not the player left the game. 0: didn't leave. 1: left safely. 2+: Abandoned */
            leaver_status?: number;
            /** Size of the players party. If not in a party, will return 1. */
            party_size?: number;
          }[];
        };
      };
    };
  };
// ...
}